+++ plugin/storage/cassandra/spanstore/dbmodel/unique_tags_test.go
@@ -17,30 +17,13 @@ package dbmodel
 import (
 	"testing"
 
-	"github.com/uber/jaeger/model"
-
 	"github.com/kr/pretty"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestFilterLogTags(t *testing.T) {
-	expectedTags := model.KeyValues{
-		model.String(someStringTagKey, someStringTagValue),
-		model.Bool(someBoolTagKey, someBoolTagValue),
-		model.Int64(someLongTagKey, someLongTagValue),
-	}
-	testSpan := getTestJaegerSpan()
-	testSpan.Tags = expectedTags
-	testSpan.Process.Tags = model.KeyValues{}
-	testSpan.Logs = []model.Log{
-		{
-			Timestamp: someLogTimestamp,
-			Fields: model.KeyValues{
-				model.Float64(someDoubleTagKey, someDoubleTagValue),
-			},
-		},
-	}
-	uniqueTags := FilterLogTags()(testSpan)
+func TestGetUniqueTags(t *testing.T) {
+	expectedTags := getTestUniqueTags()
+	uniqueTags := GetAllUniqueTags(getTestJaegerSpan(), DefaultTagFilter())
 	if !assert.EqualValues(t, expectedTags, uniqueTags) {
 		for _, diff := range pretty.Diff(expectedTags, uniqueTags) {
 			t.Log(diff)
